Update qhelp to mention solution using urlparse.
